mac os x compile fixes, ui_demoSortDirFirst, MSG_HashKey() utf8 support

+++ code/cgame/cg_draw.c
@@ -5440,7 +5440,7 @@ static float CG_DrawPowerups( float y ) {
 		powerups = cg_entities[wcg.clientNum].currentState.powerups;
 
 		if (ARRAY_LEN(powerupsCheck) >= MAX_POWERUPS) {
-			Com_Printf("^1invalid number of powerups to check: %d\n", ARRAY_LEN(powerupsCheck));
+			Com_Printf("^1invalid number of powerups to check: %d\n", (int)ARRAY_LEN(powerupsCheck));
 			return y;
 		}
